+++ CHANGELOG.md
@@ -7,6 +7,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 cluster API. See the changelog of the [cluster API](https://cluster-api.cyberfusion.nl/redoc#section/Changelog) for 
 detailed information.
 
+## [1.62.0]
+
+### Added
+
+- Add database sync endpoint.
+
+### Changed
+
+- Update Passenger app restart + FPM pool restart endpoints: add callback URL + return task collection model
+
 ## [1.61.0]
 
 ### Added